[change] Renames getFormattedTime to getFormattedUTC.
